http://192.168.37.159:3000/project/8/interface/api/430 接口翻译完成

+++ longjian-measure-app/src/main/java/com/longfor/longjian/measure/app/vo/appMeasureSyncVo/ResultListVo.java
@@ -57,15 +57,15 @@ public class ResultListVo {
     /**
      * 得分
      */
-    private Double score;
+    private Float score;
     /**
      * 更新时间
      */
-    private Integer update_at;
+    private Long update_at;
     /**
      * 删除时间(0表示未删除)
      */
-    private Integer delete_at;
+    private Long delete_at;
     /**
      * 描区uuid
      */
